Bugfix: No longer schedule highlighting StatusLine
It was causing massive CPU usage for some reason.  I thought it wouldn't
run too fast if the rest of the function were synchronous, but
apparently that doesn't matter: never schedule shit in Galaxyline.

This does introduce a frame or two of lag for the StatusLine
highlighting, but this avoids neovim running the CPU to max just for
being open.